services/list_items_generator_service/requirements.txt
anyio==3.7.1
asgiref==3.8.1
asn1crypto==0.24.0
asttokens==2.2.1
attrs==23.1.0
baarutil==1.9.2
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
Django==4.2.11
djangorestframework==3.15.1
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.103.1
filelock==3.12.0
fonttools==4.47.0
fsspec==2023.6.0
ftfy==6.1.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.1.1
ipykernel==6.23.2
ipython==8.14.0
isodate==0.6.1
jedi==0.18.2
jsonpath-ng==1.5.3
jupyter-client==8.2.0
jupyter-core==5.3.1
kiwisolver==1.4.5
lxml==4.9.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mysql-connector-python==8.1.0
nest-asyncio==1.5.6
numpy==1.24.3
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.5.3
playwright==1.34.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
pybase64==1.2.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pyee==9.0.4
Pygments==2.15.1
PyNaCl==1.5.0
pyodbc==4.0.39
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
questionary==1.10.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
robotframework==6.0.2
robotframework-crypto==0.3.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
tldextract==3.4.4
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.23.2
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
wget==3.2
win32security==2.1.0
zeep==4.2.1
zipp==3.15.0
